feat: 增加webmcp-polyfill并优化brige.ts逻辑
