+++ packages/doc-ai/src/App.vue
@@ -247,6 +247,10 @@ const mcpServers: Record<string, McpServerConfig> = {
     : {})
 }
 
+nav.modelContextTesting?.registerToolsChangedCallback((...args) => {
+  console.log('tools changed', args)
+})
+
 const menuItems = ref<any[]>([])
 
 onMounted(async () => {
