+++ packages/next-sdk/package.json
@@ -56,6 +56,7 @@
     "@ai-sdk/deepseek": "catalog:",
     "@ai-sdk/provider": "catalog:",
     "@ai-sdk/mcp": "catalog:",
+    "@mcp-b/webmcp-polyfill": "^2.2.0",
     "qrcode": "catalog:",
     "zod": "catalog:",
     "ajv": "catalog:",
